requirements.txt
# Python dependencies for DSPy integration
dspy-ai==2.5.7
google-generativeai>=0.3.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
python-dotenv==1.0.1
